pyproject.toml
[project]
name = "karuta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py-self>=2.0.1",
    "python-dotenv>=1.1.1",
]
